FontWrap! Dun dun dunnnnn!
Finally a class to take care of font stuff, specifically word wrapping
and rendering to a surface more directly. Testing it out in TestMode,
all seems good. I might add more functionality later, but this seems
good for now, might be all I need.
Basically, all of the display stuff will be making use of this.