Image magick convert files (#3)
I used the mogrify tool (running command "mogrify *.png" in each image directory) from ImageMagick-7.0.7-7-portable-Q16-x86 to convert the headings of the pngs used by the game. This is to fix a warning that would appear on image load: "libpng warning: iCCP: known incorrect sRGB profile".

Fixing this caused a bug for some combinations of monster sprites, I fixed this by calling .convert_alpha() on them so the transparency is handled correctly.