Move shared stuff out of static (#6)
* moving shared stuff around
* moving shared stuff around
* moving shared stuff around
* init
* monster clean up and shared state
* realized that MonConvoMode needs to be sure to draw the right image
* conversions onto the screen they'll actually draw on
* putting the directory joining in the constants file, simpler
* make boxes a little simpler, clean up fight mode
* cleaning up convomode as well
* saving my spot for now
* hmm
* wait, use a module. still have to separate shared and state in other files and put other display stuff in shared
* gamemenumode and game cutover
* using shared
* splitting files to hold single classes, fixing imports to fit PEP 8
* might be close
* almost
* fixing problems introduce by find and replace
* doesn't matter

4 days of work, but this makes more sense I think!

Instead of each gamemode child having access to a shared variable for shared state, just putting that in a proper module makes a lot more sense (for singleton state). More organized, able to ensure it's initialized before everything that needs it (thanks import!), no need to worry about making it too many times (thanks again import!).

Also I reorganized how the Boxes class is used. It can hold it's own state and just be an instance, much simpler.

Also clean up all my imports, I think they're all fitting recommended style now? As part of that, separated out some files so as to not have files with more than one class definition.

Also sharedstate > shared and state. The former holds game stuff that everything needs access to that's not a simple constant, or is directly linked to something that's not a simple constant, and also has functions for manipulating those variables. The latter holds state that's part of the game but shared between modes (something that will need to be savable).

A lot of assorted cleanup happened as I went, often with a mind towards making saving the game easier in the future.

started moving some classes to using slots instead of the default dict. This reduces memory usage, and will continue.